+++ dio/lib/src/utils.dart
@@ -5,14 +5,6 @@ import 'dart:convert';
 import 'options.dart';
 import 'parameter.dart';
 
-/// A regular expression that matches strings that are composed entirely of
-/// ASCII-compatible characters.
-final RegExp _asciiOnly = RegExp(r'^[\x00-\x7F]+$');
-
-/// Returns whether [string] is composed entirely of ASCII-compatible
-/// characters.
-bool isPlainAscii(String string) => _asciiOnly.hasMatch(string);
-
 /// Pipes all data and errors from [stream] into [sink]. Completes [Future] once
 /// [stream] is done. Unlike [store], [sink] remains open after [stream] is
 /// done.